IIB project

+++ Scripts/ZmqServer.cs
@@ -38,7 +38,7 @@ namespace UserInTheBox
         public bool isFinished;
         public float reward;
         public byte[] image;
-        public byte[] audio;  // new field for audio data
+        public float[] audio;  // new field for audio data
         public float timeFeature;
         public string logDict;  // json dict
     }
@@ -115,13 +115,13 @@ namespace UserInTheBox
             state = JsonUtility.FromJson<TMessage>(strMessage);
         }
 
-        public void SendObservation(bool isFinished, float reward, byte[] image, byte[] audio, float timeFeature, Dictionary<string, object> logDict)
+        public void SendObservation(bool isFinished, float reward, byte[] image, float[] audio, float timeFeature, Dictionary<string, object> logDict)
         {
             // Populate reply
             _gameObservation.isFinished = isFinished;
             _gameObservation.reward = reward;
             _gameObservation.image = image;
-            _gameObservation.audio = audio;  // set audio data
+            _gameObservation.audio = audio;
             _gameObservation.timeFeature = timeFeature;
             _gameObservation.logDict = JsonConvert.SerializeObject(logDict);
